+++ tests/assets/conf.py
@@ -194,6 +194,7 @@ epub_exclude_files = ["search.html"]
 # -- Extension configuration -------------------------------------------------
 autodoc_default_options = {
     "private-members": True,
+    "undoc-members": True,
 }
 
 primary_domain = "c"